Add functionality to create & delete IAM role
An IAM role in the attacker-controlled AWS is required to be able to perform enumeration.

This change adds functionality to create and delete that role.

The role name can be specified with the `-r` CLI parameter, with a default set to `IAMEnum` .

Te role name is suffixed with an 8 character random string to avoid naming conflicts.